Create separate module for borer utilities
Moving common code working with borer library from solr-client and
commons to its own module to be used in other library modules.
