+++ modules/json/src/main/scala/io/renku/json/EncoderSupport.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.renku.solr.client
+package io.renku.json
 
 import scala.compiletime.*
 import scala.deriving.*
